clean up and update library user guide

+++ src/include/pkt_proc.h
@@ -53,7 +53,7 @@
 void* process_packet(unsigned char *ctx_ptr, const struct pcap_pkthdr *header, const unsigned char *packet);
 void libpcap_process_packet(unsigned char *ctx_ptr, const struct pcap_pkthdr *header, const unsigned char *packet);
 
-int get_packet_5tuple_key(const unsigned char *packet, flow_key_t *key);
+uint8_t get_packet_5tuple_key(const unsigned char *packet, flow_key_t *key);
 
 joy_status_e process_ipfix(joy_ctx_data *ctx, const char *start, int len, flow_record_t *r);
